requirements.txt
torch>=2.1.0
transformers>=4.45.0
peft>=0.12.0
trl>=0.11.0
bitsandbytes>=0.43.0
accelerate>=0.34.0
datasets>=3.0.0
pyyaml>=6.0
openai>=1.0.0
anthropic>=0.34.0
wandb>=0.17.0
